+++ src/hooks/useBlueprintCalculation.ts
@@ -9,7 +9,8 @@ export const useBlueprintCalculation = (
   imageRef: RefObject<HTMLImageElement>,
   size: Size,
   config: Config,
-  setBlueprint: (blueprint: FactorioBlueprint) => void
+  setBlueprint: (blueprint: FactorioBlueprint) => void,
+  setPreviewSrc: (url: string) => void
 ) => {
   useEffect(() => {
     const curImage = imageRef.current;
@@ -22,9 +23,10 @@ export const useBlueprintCalculation = (
         size.height
       ).data;
 
-      calculateBlueprint(data, size, config).then((b) => {
+      calculateBlueprint(data, size, config).then(([b, preview]) => {
         setBlueprint(b);
+        setPreviewSrc(preview);
       });
     }
-  }, [config, imageRef, setBlueprint, size]);
+  }, [config, imageRef, setBlueprint, setPreviewSrc, size]);
 };